Payments matched to unique booking id instead of film name
In the initial code, save_payment would create a payment without a film
if there were more than one booking of the same film. Since it is
entirely possible to book a film more than once, the function now
receives the id from LinkCol in the Bookings table. This ensures a
payment is made to that unique booking, regardless of film.