Create 1:M relationship between organizations & distributors
Distributors are now created on a per-organization basis since not
every organization works with the same distributors. If the original
logic was kept, then if a user from Org A edits a distributor, it
changes for Org B. It also ran the risk of exposing contact info for
distributors only certain organizations use.

The downside to the new approach is there will be lots of duplicate
entries depending on how many organizations use the service.