Change login to require email instead of username
It's much easier for users to remember their email than another username 
they had to create for the service. This does beg the question of 
whether usernames are at all necessary for this service. There should 
never be a need for a user to see users from an organization they do not 
belong to.